+++ src/anomalib/data/datasets/__init__.py
@@ -42,6 +42,7 @@ Example:
 from .base import AnomalibDataset, AnomalibDepthDataset, AnomalibVideoDataset
 from .depth import Folder3DDataset, MVTec3DDataset
 from .image import (
+    BMADDataset,
     BTechDataset,
     DatumaroDataset,
     FolderDataset,
@@ -63,6 +64,7 @@ __all__ = [
     "Folder3DDataset",
     "MVTec3DDataset",
     # Image
+    "BMADDataset",
     "BTechDataset",
     "DatumaroDataset",
     "FolderDataset",