+++ CHANGELOG.md
@@ -8,6 +8,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Added
 
+- 🚀 Add BMAD dataset by @code-dev05 in https://github.com/open-edge-platform/anomalib/pull/2900
+
 ### Removed
 
 ### Changed